fix: prohibit the 'retry' and 'retry_timeout' opts
If the caller sets the `retry > 0` when starting the ehttpc,
the down connections will be started by the underlying gun client
periodically.

So if one calls the health_check when `gun_state=down`,
ehttpc will leak ports because it opens an new TCP connection, but
the old connection was not closed.

We forcely set `retry = 0`, and prohibit the user from setting
the two opts: `retry` and `retry_timeout`.