fix: crash and no message send with body headers and no body
When one sent a message with, for example, the content-type header and
no body, the underlying gun library would expect a body to come with
a gun:data/3 call and would not finish the request. The next request
would then crash. See the following issue for more information:

https://github.com/ninenines/gun/issues/141

To write regression testcases for this issue, I also had to make changes
to the test server so that it would handle the content type
content-length.

Fixes:
https://emqx.atlassian.net/browse/EMQX-8980