fix(pool): forcefully recreate pool if already exists
There are (extremely rare) situations when a pool that is under control
of the `gproc_pool` server may end up in an inconsistent state. This
is possible because `gproc_pool:new/3` involves a sequence of operations
against `gproc` state, and this sequence is not atomic. In extreme
conditions, one of the operations may fail, kill the `gproc_pool` server
and leave the pool state inconsistent. The reasonable (and advertised)
way to deal with it is to `force_delete` it first. This should be safe
as the assumption is that an `ehttpc_pool` owns its `gproc_pool`.